requirements.txt
Flask==3.0.3
flask-mail==0.9.1
python-dotenv==1.0.1
supabase==2.3.4
werkzeug==3.0.4
requests==2.32.3
httpx==0.24.1